requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
djangorestframework==3.14.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
